LPS-67652 Move upgrade process from 7002 to 7003 as they weren't included in released 7.0.2 GA3
