+++ portal-impl/src/portal.properties
@@ -132,7 +132,7 @@
         com.liferay.portal.upgrade.UpgradeProcess_6_2_0\,\
         com.liferay.portal.upgrade.UpgradeProcess_7_0_0\,\
         com.liferay.portal.upgrade.UpgradeProcess_7_0_1\,\
-        com.liferay.portal.upgrade.UpgradeProcess_7_0_2
+        com.liferay.portal.upgrade.UpgradeProcess_7_0_3
 
     upgrade.processes.6012=${upgrade.processes.master}
     upgrade.processes.6100=${upgrade.processes.master}
