+++ portal-kernel/src/com/liferay/portal/kernel/util/ReleaseInfo.java
@@ -99,6 +99,8 @@ public class ReleaseInfo {
 
 	public static final int RELEASE_7_0_2_BUILD_NUMBER = 7002;
 
+	public static final int RELEASE_7_0_3_BUILD_NUMBER = 7003;
+
 	public static final int RELEASE_7_0_10_BUILD_NUMBER = 7010;
 
 	public static final Date getBuildDate() {
